Rename "testcases" namespace to "clinicaltrials".
