+++ resources/testng-CTS.xml
@@ -9,10 +9,13 @@
 			</run>
 		</groups>
 		<classes> 	
-			  <class name="gov.nci.testcases.BasicSearch_Test" />  
-			   <!-- <class name="gov.nci.testcases.AdvanceSearch_Test" /> 
-			    <class name="gov.nci.testcases.AdvanceSearchResults_Test" />    		
-			<class name="gov.nci.testcases.BasicSearchResults_Test" />-->
+			  <class name="gov.nci.clinicaltrials.BasicSearch_Test">
+			  	<methods>
+				</methods>
+			  </class>
+			   <!-- <class name="gov.nci.clinicaltrials.AdvanceSearch_Test" /> 
+			    <class name="gov.nci.clinicaltrials.AdvanceSearchResults_Test" />    		
+			<class name="gov.nci.testcclinicaltrialsases.BasicSearchResults_Test" />-->
 			    
 		</classes>
 	</test> 
@@ -25,11 +28,11 @@
 			</run>
 		</groups>
 		<classes> 	
-			 <class name="gov.nci.testcases.BasicSearch_Test" /> 
-			<class name="gov.nci.testcases.AdvanceSearch_Test" />
+			 <class name="gov.nci.clinicaltrials.BasicSearch_Test" /> 
+			<class name="gov.nci.clinicaltrials.AdvanceSearch_Test" />
 			
 			 
-			  <class name="gov.nci.testcases.AdvanceSearchResults_Test" />  
+			  <class name="gov.nci.clinicaltrials.AdvanceSearchResults_Test" />  
 			
 		</classes>
 	</test>  -->
@@ -42,10 +45,10 @@
 			</run>
 		</groups>
 		<classes> 	
-		 <class name="gov.nci.testcases.BasicSearch_Test" /> 
-		<class name="gov.nci.testcases.AdvanceSearch_Test" />
+		 <class name="gov.nci.clinicaltrials.BasicSearch_Test" /> 
+		<class name="gov.nci.clinicaltrials.AdvanceSearch_Test" />
 			
-		 <class name="gov.nci.testcases.AdvanceSearchResults_Test" /> 
+		 <class name="gov.nci.clinicaltrials.AdvanceSearchResults_Test" /> 
 			
 		</classes>
 	</test> -->
@@ -57,8 +60,8 @@
 			</run>
 		</groups>
 		<classes> 	
-			<class name="gov.nci.testcases.BasicSearch_Test" />
-		<class name="gov.nci.testcases.AdvanceSearch_Test" />
+			<class name="gov.nci.clinicaltrials.BasicSearch_Test" />
+		<class name="gov.nci.clinicaltrials.AdvanceSearch_Test" />
 		</classes>
 	</test>  -->
 	<!--   <test name="NCI - Firefox Testing" >
@@ -69,8 +72,8 @@
 			</run>
 		</groups>
 		<classes> 	
-			 <class name="gov.nci.testcases.BasicSearch_Test" /> 
-			   <class name="gov.nci.testcases.AdvanceSearch_Test" />
+			 <class name="gov.nci.clinicaltrials.BasicSearch_Test" /> 
+			   <class name="gov.nci.clinicaltrials.AdvanceSearch_Test" />
 		</classes>
 	</test>  --> 
 	<!-- <test name="NCI - Firefox Testing Headless" >
@@ -81,8 +84,8 @@
 			</run>
 		</groups>
 		<classes> 	
-			<class name="gov.nci.testcases.BasicSearch_Test" />
-			  <class name="gov.nci.testcases.AdvanceSearch_Test" /> 
+			<class name="gov.nci.clinicaltrials.BasicSearch_Test" />
+			  <class name="gov.nci.clinicaltrials.AdvanceSearch_Test" /> 
 		</classes>
 	</test>   -->
 </suite>
