+++ resources/testng-R4R.xml
@@ -9,12 +9,12 @@
 			</run>
 		</groups>
 		<classes> 	
-			 <class name="gov.nci.testcases.BasicSearch_Test" /> 
+			 <class name="gov.nci.clinicaltrials.BasicSearch_Test" /> 
 			 
 			    
 			  <!--    	
-			  <class name="gov.nci.testcases.AdvanceSearch_Test" />
-			  <class name="gov.nci.testcases.AdvanceSearchResults_Test" />  --> 	
+			  <class name="gov.nci.clinicaltrials.AdvanceSearch_Test" />
+			  <class name="gov.nci.clinicaltrials.AdvanceSearchResults_Test" />  --> 	
 
 		</classes>
 	</test> 
@@ -27,11 +27,11 @@
 			</run>
 		</groups>
 		<classes> 	
-			 <class name="gov.nci.testcases.BasicSearch_Test" /> 
-			<class name="gov.nci.testcases.AdvanceSearch_Test" />
+			 <class name="gov.nci.clinicaltrials.BasicSearch_Test" /> 
+			<class name="gov.nci.clinicaltrials.AdvanceSearch_Test" />
 			
 			 
-			  <class name="gov.nci.testcases.AdvanceSearchResults_Test" />  
+			  <class name="gov.nci.clinicaltrials.AdvanceSearchResults_Test" />  
 			
 		</classes>
 	</test>  -->
@@ -44,10 +44,10 @@
 			</run>
 		</groups>
 		<classes> 	
-		 <class name="gov.nci.testcases.BasicSearch_Test" /> 
-		<class name="gov.nci.testcases.AdvanceSearch_Test" />
+		 <class name="gov.nci.clinicaltrials.BasicSearch_Test" /> 
+		<class name="gov.nci.clinicaltrials.AdvanceSearch_Test" />
 			
-		 <class name="gov.nci.testcases.AdvanceSearchResults_Test" /> 
+		 <class name="gov.nci.clinicaltrials.AdvanceSearchResults_Test" /> 
 			
 		</classes>
 	</test> -->
@@ -59,8 +59,8 @@
 			</run>
 		</groups>
 		<classes> 	
-			<class name="gov.nci.testcases.BasicSearch_Test" />
-		<class name="gov.nci.testcases.AdvanceSearch_Test" />
+			<class name="gov.nci.clinicaltrials.BasicSearch_Test" />
+		<class name="gov.nci.clinicaltrials.AdvanceSearch_Test" />
 		</classes>
 	</test>  -->
 	<!--   <test name="NCI - Firefox Testing" >
@@ -71,8 +71,8 @@
 			</run>
 		</groups>
 		<classes> 	
-			 <class name="gov.nci.testcases.BasicSearch_Test" /> 
-			   <class name="gov.nci.testcases.AdvanceSearch_Test" />
+			 <class name="gov.nci.clinicaltrials.BasicSearch_Test" /> 
+			   <class name="gov.nci.clinicaltrials.AdvanceSearch_Test" />
 		</classes>
 	</test>  --> 
 	<!-- <test name="NCI - Firefox Testing Headless" >
@@ -83,8 +83,8 @@
 			</run>
 		</groups>
 		<classes> 	
-			<class name="gov.nci.testcases.BasicSearch_Test" />
-			  <class name="gov.nci.testcases.AdvanceSearch_Test" /> 
+			<class name="gov.nci.clinicaltrials.BasicSearch_Test" />
+			  <class name="gov.nci.clinicaltrials.AdvanceSearch_Test" /> 
 		</classes>
 	</test>   -->
 </suite>
