+++ src/test/java/gov/nci/clinicaltrials/AdvanceSearch_Test.java
@@ -1,4 +1,4 @@
-package gov.nci.testcases;
+package gov.nci.clinicaltrials;
 
 import java.net.MalformedURLException;
 import java.util.ArrayList;
