+++ src/test/java/gov/nci/clinicaltrials/BasicSearch_Test.java
@@ -1,4 +1,4 @@
-package gov.nci.testcases;
+package gov.nci.clinicaltrials;
 
 import java.net.URL;
 import java.io.UnsupportedEncodingException;
